Refactoring model and component file names
