+++ backend/routes/users.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 const bcrypt = require('bcrypt');
-let User = require('../models/user.model');
+let User = require('../models/user');
 
 router.route('/').get((req, res) => {
     User.find()
