+++ client/src/components/createUser.js
@@ -54,7 +54,7 @@ export default class CreateUser extends Component {
                 <form onSubmit={this.onSubmit}>
                     <div className="form-group">
                         <label>Email address</label>
-                        <input type="text"
+                        <input type="email"
                             required
                             className="form-control"
                             value={this.state.email}
